Dedup child view imports
Remove duplicate imports when importing child views.

Duplicate imports can happen if a view has two or more
child views with the same name.

Using the same name should generally be avoided but it
can happen when using symbols on Webflow in which case
it might be wanted since the rest of the content of
the symbol might also be identical.