Remove default proxies if there's no controller
The default proxies are not needed because the value of proxies[name]
is passed to a helper map() function which checks if it's undefined.

They also prevent the view from being used directly, without a
controller.

Also, they were only defined for root proxies, not for nested ones
(nested ones were actually included as root and never used).