Add support for page folders
Pages can be placed in folders in Webflow.

Before this change, those pages were ignored and simply
copied into the "public" directory.

After this change, files in the input ".appfairy" directory (and all
its subdirectories) are divided into HTML files and "other" files.

Other files are copied into the "public" folder and modified
in-place as needed (css, for example).

HTML files are transpiled into views and written into:
src/views/[folder/subfolder]/SomeView.js

Controllers are expected to be in the corresponding location:
src/controllers/[folder/subfolder]/SomeController.js

Child views (elements with af-el="name" attribute) are written
into the same directory as the parent view.

src/views/index.js only includes the root views and their childs.

This change makes it possible to create a namespace for a page and
its children by placing a page in a folder.