create a state file instead of looking at git
Appfairy locates the files created by previous migration by
looking for a git commit it created in the past.

This has several issues:
* The migration commits can't be squashed with other commits
* There can be only one project per git repo (one git history)
* Not all files are located because only the last commit is checked

This change replaces that with a state file that is created in the
output directory during the migration and which contains a list of
all files auto-generated by Appfairy.

This file should be checked into git and is used during the next
migration to locate and remove the old migration files before the
new files are generated.

This way, files with new names (for example after a page has been
renamed in Webflow) don't end up duplicated.

All changes are still staged (git add) after the migration is
complete but commiting is left to the user.