input is now a zip file
Exporting from Webflow provides a zip file so having to extract it
into the .appfairy directory was an unnecessary extra step (that
could go wrong if the directory already existed).

Now Appfairy works off of the zip file directly, decompressing it
on the fly (no temporary directory).

The name of the file can be provided using the --input argument,
the default is "webflow.zip" in the root of the project.