add alternative proxy style
Standard style:

<FooView>
  <title>My View</title>
</FooView>

Alternative style:

<FooView>
  <proxy af-sock="title">My View</proxy>
</FooView>

The tag name ("proxy" in this example) is not used and can be anything.
In practice, it might be a good idea to use the same html element as
used by the view:

<FooView>
  <a af-sock="about-link" href="/about" />
</FooView>

Motivation:

When using TypeScript to write controllers, the compiler complains
about proxy elements being undefined ("title" in the standard-style
example above).