add af-repeat attribute
Add a new attribute that can be specified along with af-sock
in Webflow.

af-repeat="<repeat-mode>"

The repeat mode controlls how the element is rendered for a given
number of proxies in the controller. Whether a proxy needs to
be specified at all and whether it can be repeated more than once.

There are 4 repeat modes and they resemble the regex repetition
operators.

af-repeat=""
------------

This is the default mode, also used when the attribute is not set.

The output will have exactly one element irrespective of how many
proxies are in the controller.

0 proxies => render Webflow element unchanged
1 proxy => render the proxy
2 or more => throw a runtime error

af-repeat="?"
-------------

The output will have 0 or 1 element depending on the number of
proxies in the controller.

0 proxies => skip the element
1 proxy => render the proxy
2 or more => throw a runtime error

af-repeat="+"
-------------

The output will have 1 or more elements depending on the number
of proxies in the controller.

0 proxies => render Webflow element unchanged
1 or more => render the proxies as siblings

af-repeat="*"
-------------

The output will have 0 or more elements depending on the number
of proxies in the controller.

0 proxies => skip the element
1 or more => render the proxies as siblings

---

Previously all sockets behaved as if they had af-repeat="*" which
meant that they had to be included in the controller to be rendered
at all.

The new default repeat renders them without having to add the proxies
while at the same time it makes sure that elements that are really
meant to be repeated are marked as such (using "+" or "*").