move script handling to viewHelpers
Move some script-handling code out of the view file and into
the viewHelpers.js file so it can be shared by all views.

View helpers includes a shared store for scripts so they
don't have to be fetched multiple times.

It also includes a new reinitWebflow() function to re-initialize
the Webflow event handlers which is useful when working with triggers
and animations in Webflow. It also uses the shared store.