make css encapsulation optional, off by default
The encapsulation, as it is currently implemented, is quite
fragile, not to say "hacky", and not immune to future changes
to how styling works in Webflow.

At the same time, other than trying to keep the styling local
to an Appfairy-view (for cases where they're used as components
in other designs), there's little reason for it.

This change turns it into an optional feature.

The encapsulation of normalize.css, which was disabled before,
is now also controlled by this setting.