redesign sub-views
The af-el="nav-bar" attribute is now called af-view="nav-bar"
and works differently.

The sub-views are now generated inside the page-view file (no
longer in separate files) in namespaces corresponding to their
location on the page. For example:

IndexView.NavBarView.LoginView

Conflicts can only happen if there are two or more views with
the same name on one page and at the same depth, which should
be rare. If it happens, the migration is aborted.

Every view has a separate "sock" enum accessible via FooView.sock
with just the sockets used by that view. The "sock" enum of the
page view is still also available at module level for convenience.

Controller logic has been removed entirely - Appfairy makes no
assumptions about where controllers are (they don't even have to be
in "controllers/" anymore). Accessing them via <FooView.Controller />
is no longer supported. Import and use them directly instead.

To assign a controller to a sub-view, attach a socket to it and
pass your controller to the proxy:

<IndexView>
    <MyNavBarController af-sock={sock.nav_bar} />
</IndexView>

Any other props are passed on to the controller.

When using sub-views in reusable Webflow components, a view will be
generated for every component instance but they should all be using
a single controller.

To achieve that, the controller can be reused in multiple
proxies but it needs to be told which view to use.

To make this easier, the proxy automatically passes the correct
view to the controller via a prop.

export const MyNavBarController = (props) => {
    return (
        <props.NavBarView>
        </props.NavBarView>
    )
}

The name of the prop is the same as the name of the view.