provide default af-sock for af-view elements
A view with no socket is not very useful because there's no way
to attach a controller to it and therefore no way to access its
own sockets.

Most of the time the value of af-sock is expected to be the same
as the value of af-view so it makes sense to reuse af-view as a
default for af-sock.