allow proxies to override tag types
First, this removes the old style of proxies.

<mylink href="" />

From now on, only the new syntax works.

<a af-sock="mylink" href="" />

The tag type of the proxy (<a>) will now also override
the type from Webflow.

This brings the normal proxies in sync with child view proxies,
which were already using the tag type to specify the controller
for the child view.