use the new scripts loader in views
This means that webflow.js finally isn't loaded twice which is
what was causing the "u not defined"-type of errors at runtime.

The new loader adds <script> elements to the <head> instead
of fetching and eval()-ing the scripts.

Move the list of view scripts into the view class.

Add a base class (View) for all views.

The base class takes care of scripts loading and Webflow ix2
updates which makes sure that Webflow event handlers are correctly
attached and "Page load" events are triggered.

This means that the reinitWebflow() function is no longer needed.